backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
sqlalchemy[asyncio]
python-dotenv
pydantic-settings
alembic
